Moved the import functions from expr_visitor to stmt_visitor
